requirements.txt
mysql-connector-python==9.4.0
prettytable==3.16.0
python-dotenv==1.0.0